pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grounding-ai"
version = "0.3.0"
description = "Local-first document corpus pipeline for grounded AI agents"
authors = [{name = "Andy"}]
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "MIT"}

dependencies = [
    "unstructured>=0.15.0",
    "pdfminer.six>=20231228",
    "langchain-text-splitters>=0.2.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "ujson>=5.10",
    "blake3>=0.4.1",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.0",
    "music21>=9.1.0",
    "pillow>=10.0.0",
    "mido>=1.2.0",
    "pix2tex>=0.1.2",
    "pypdfium2>=4.0.0",
    "scipy>=1.10.0",
    "typer>=0.9.0",
    "rank_bm25>=0.2.2",
]

[project.scripts]
grounding = "grounding.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["grounding*"]
